moving web_views to web_interface/page_views
