+++ docs/release_notes/archived/0_7_8.md
@@ -29,7 +29,7 @@ from sageworks.views.artifacts_view import ArtifactsView
 is now
 
 ```
-from sageworks.web_views.artifacts_web_view import ArtifactsWebView
+from sageworks.web_interface.page_views.artifacts_web_view import ArtifactsWebView
 ```
 
 ## Questions?
