+++ src/sageworks/utils/plugin_manager.py
@@ -12,7 +12,7 @@ from typing import Union, Dict, List, Any
 # SageWorks Imports
 from sageworks.utils.config_manager import ConfigManager
 from sageworks.utils.s3_utils import copy_s3_files_to_local
-from sageworks.web_views.web_view import WebView
+from sageworks.web_interface.page_views.web_view import WebView
 from sageworks.web_components.plugin_interface import PluginInterface
 from sageworks.web_components.plugin_interface import PluginPage
