+++ src/sageworks/web_interface/page_views/artifacts_text_view.py
@@ -4,7 +4,7 @@ import pandas as pd
 from typing import Dict
 
 # SageWorks Imports
-from sageworks.web_views.web_view import WebView
+from sageworks.web_interface.page_views.web_view import WebView
 from sageworks.cached.cached_meta import CachedMeta
