+++ src/sageworks/web_interface/page_views/feature_set_web_view.py
@@ -3,7 +3,7 @@
 import pandas as pd
 
 # SageWorks Imports
-from sageworks.web_views.artifacts_web_view import ArtifactsWebView
+from sageworks.web_interface.page_views.artifacts_web_view import ArtifactsWebView
 from sageworks.cached.cached_feature_set import CachedFeatureSet
